@@ -1,7 +1,17 @@
 apply plugin: 'com.android.application'
+apply from: '../dependencies.gradle'
+apply from: '../spotless.gradle'
 
 android {
-    compileSdkVersion 28
+    signingConfigs {
+        release {
+            storeFile file('/home/dmj/Documents/Clients/Direct - OlavAndMic/keystore/original keystore')
+            storePassword password
+            keyAlias = 'll'
+            keyPassword password
+        }
+    }
+    compileSdkVersion 29
 
     /*signingConfigs{
         key{
