@@ -1,6 +1,6 @@
 package com.olav.logolicious.util;
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 /**
  * Created by ASUS on 1/29/2018.
